refactor: 도메인 이름 변경 및 리팩토링
- Font라는 개체를 중심으로 설계하기로 판단
- 컨트롤러에 있던 비즈니스 로직을 service 단으로 뺌
